Combining classes, grouping them into a new directory
